extract servlet configuration methods to pkg local class.
