+++ src/main/java/com/auth0/Auth0RedirectServlet.java
@@ -11,7 +11,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-import static com.auth0.lib.ServletUtils.readLocalRequiredParameter;
+import static com.auth0.ConfigUtils.readLocalRequiredParameter;
 
 /**
  * The Servlet endpoint used as the callback handler in the OAuth 2.0 authorization code grant flow.
