+++ src/test/java/com.auth0/lib/TokensTest.java
@@ -1,6 +1,5 @@
 package com.auth0.lib;
 
-import com.auth0.lib.Tokens;
 import org.junit.Test;
 
 import static org.hamcrest.core.Is.is;
